requirements/testing.txt
pytest>=5.3.5
pytest-django>=3.8.0
pytest-cov>=2.8.1
flake8==6.0.0
mypy>=0.770
django-stubs>=1.8.0,<1.10.0
djangorestframework-stubs>=1.1.0
types-PyYAML>=0.1.6
isort>=5.0.4
autopep8==2.0.0
